pyproject.toml
[tool.poetry]
name = "gb-flask-blog"
version = "0.1.0"
description = ""
authors = ["Sergey Karnaukhov <s_karn_v@mail.ru>"]
readme = "README.md"
packages = [{include = "gb_flask_blog"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.1"
httpie = "^3.2.1"
faker = "^18.6.0"
flask-sqlalchemy = "^3.0.3"
python-dotenv = "^1.0.0"
flask-login = "^0.6.2"
flask-migrate = "^4.0.4"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
